Refactorisation of part of web/files into a vfs package
